chore: change bundler to vite
