+++ package.json
@@ -18,10 +18,15 @@
     "eslint": "^8.37.0",
     "html-webpack-plugin": "^5.5.0",
     "mini-css-extract-plugin": "^2.6.1",
-    "sass": "^1.56.0",
+    "sass": "^1.77.6",
     "sass-loader": "^13.1.0",
+    "vite": "^5.3.3",
+    "vite-plugin-static-copy": "^1.0.6",
     "webpack": "^5.74.0",
     "webpack-cli": "^4.10.0",
     "webpack-dev-server": "^4.11.1"
+  },
+  "dependencies": {
+    "@fortawesome/fontawesome-free": "^6.5.2"
   }
 }
