+++ src/index.html
@@ -15,7 +15,7 @@
       rel="stylesheet"
     />
     <title>Where are you going?</title>
-    <link rel="stylesheet" href="vendor/fontawesome/css/all.css" />
+    <link rel="stylesheet" href="./scss/main.scss" />
   </head>
   <body>
     <div class="background">
@@ -38,7 +38,6 @@
               autocomplete="off"
               id="name"
             />
-
             <br /><br />
             <div class="background-settings">
               <div class="text">
@@ -69,7 +68,6 @@
                 ><br />
               </div>
             </div>
-
             <br /><br />
             Hours Format :
             <select name="" id="myclock">
@@ -111,5 +109,7 @@
       </div>
       <footer></footer>
     </div>
+
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
