+++ src/js/background.js
