+++ src/scss/main.scss
@@ -1,3 +1,5 @@
+@import "@fortawesome/fontawesome-free/css/all.min.css";
+
 * {
   margin: 0 auto;
   padding: 0 auto;
